minor fixes

+++ commands/moderation/unmute.js
@@ -4,6 +4,7 @@ const ms = require('ms')
 module.exports = {
     name:"unmute",
         async execute(client, msg, args) {
+            if(!msg.member.hasPermission('MANAGE_MESSAGES')) return msg.reply("You do not have permission to use this command");
             msg.delete()
         var target = msg.mentions.users.first() || msg.guild.members.cache.get(args[0]);
         if(!target) return msg.reply('you need to mention a user for me to unmute them!')
